refactor: moved non-component constants to separate files
Not having these separated results in fast-refresh not being executed properly by vite
