+++ src/forms/auth/forms.jsx
@@ -1,6 +1,5 @@
 import { LoaderCircle } from 'lucide-react';
 import { Button } from '@/components/ui/button';
-import { z } from 'zod';
 import {
   Form,
   FormControl,
@@ -11,16 +10,6 @@ import {
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { PasswordInput } from '@/components/forms/password-input';
-
-export const loginSchema = z.object({
-  email: z
-    .string({ required_error: 'Email is required' })
-    .email('Invalid email address')
-    .transform((val) => val.trim()),
-  password: z
-    .string({ required_error: 'Password is required' })
-    .transform((val) => val.trim()),
-});
 
 export function LoginForm({ form, onSubmit, isLoading, error }) {
   return (
