+++ src/hooks/use-theme.jsx
@@ -1,0 +1,18 @@
+import { createContext, useContext } from 'react';
+
+const initialState = {
+  theme: 'system',
+  setTheme: () => null,
+};
+
+export const ThemeProviderContext = createContext(initialState);
+
+export function useTheme() {
+  const context = useContext(ThemeProviderContext);
+
+  if (context === undefined) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+
+  return context;
+};
